[aml-sdmmc] Fix minor bug in new tuning method and add a test
During tuning, the final window of delay line results needs to be
checked to see if it is the new largest window. This doesn't apply to
the old tuning method and likely doesn't affect any devices with new
tuning enabled, as the last failing window is typically smaller than the
overall largest one.

Bug: 68436
Test: fx test aml-sdmmc-test
Change-Id: I41049a5620c372eff679ab4d9fe7d0624023f70a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605674
Commit-Queue: Braden Kell <bradenkell@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>