[llcpp] Remove arena parameter for table fields <= 4 bytes in size
This CL removes the table field setter for fields <= 4 bytes in size
that takes an arena as a parameter. This field was deprecated as part of
the move to wire format v2 and the shift to inline field setters where
possible.

Bug: 79578
Change-Id: I495ee474930866cfe380f298a3dd8418007682d7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601782
Fuchsia-Auto-Submit: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Pascal-Louis Perez <pascallouis@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>