[usb][xhci] fix locking
Originally we were obtaining the lock in a function, release it, and
then call a sub-function and obtain the lock again. This allows for a
potential race where another thread could acquire the lock before we
re-acquire it in the sub-function.

This CL requires the sub-function to be called when holding the lock,
eliminating the possibility of race conditions.

Test: tested on NUC

Change-Id: I170960f75fc39a1e55ef25b2627c24a0d47c0086
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/608508
Commit-Queue: Ruby Zhuang <rdzhuang@google.com>
Reviewed-by: Sam Hansen <hansens@google.com>