[usb][xhci] fix new segment logic and trbs count
A newly added segment in the event ring (when the event rings grow)
could or could not be used in the current cycle depending on the
position of EREP. Originally we were not dealing with this logic. This
change adds this logic according to 4.9.4.1 in the spec

We are also only incrementing the TRBs counter, but never decrementing.
In this change, we decrement the TRBs counter when we process one TRB
(through the advancement of ERDP). Without this, new segments are too
frequent and causes lag in between mouse movements.

Bug: 81262

Test: Mouse working on NUC for 4 hours

Change-Id: I6b6ed92a0792b11f8de933ee1ab93d1301153e80
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/608509
Commit-Queue: Ruby Zhuang <rdzhuang@google.com>
Reviewed-by: Sam Hansen <hansens@google.com>