[cros-ec][lpc] Add ChromeOS EC LPC interface driver.
This is phase 1 of moving the ChromeOS EC drivers out of the x86 board
driver. This CL implements the transport protocol for EC drivers. A
future CL will add a driver that binds to this driver and communicates
with the EC to provide functionality to the rest of the system.

Bug: 78349
Change-Id: Id495d0c320187190a43642d9fce5f4e4f9d61fef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/608118
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
API-Review: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Simon Shields <simonshields@google.com>