[realm_builder][rust] rename ChildProperties to ChildOptions
The ChildProperties struct was introduced to bring the Rust client
library closer to the API outlined in RFC-0140, but this landed before
the RFC did and the name was then changed late in the RFC process.

Now that RFC-0140 has landed, update the name in the Realm Builder Rust
client library to match the finalized API names.

Change-Id: I7752c93dfd51f0e2f43d14e1024d8b1f89f3f400
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610282
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Yaneury Fermin <yaneury@google.com>
Commit-Queue: Derek Gonyeo <dgonyeo@google.com>