[llcpp] Add Transport parameter to BindServer and related functions
After this CL goes in, it should be possible to define servers for
alternative transports.

While there are no tests in this CL, building with an alternative
transport was tested with an in-progress CL introducing an integration
test.

Bug: 85734
Change-Id: Iae0645edb63b1d9f4a01bfb8ff524f77174d8060
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/606041
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
API-Review: Pascal Perez <pascallouis@google.com>