[fdf][platform] Change acpi binding parent
Devices which add children manually must
be first added with the DEVICE_ADD_NON_BINDABLE
flag set.
The acpi device builder was adding children to
devices that needed to be bindable.  To fix this
all acpi devices are now added to the acpi root,
which is given the flag DEVICE_ADD_NON_BINDABLE.

Bug: 84362
Test: device enumeration test.

Change-Id: I8fb14cae5e7bbbb05bb5ef58d8272336558c05fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/608845
Commit-Queue: Garratt Gallagher <garratt@google.com>
Reviewed-by: Simon Shields <simonshields@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>