[fvm] Convert partition limits to slices.
Previously the FVM FIDL API used bytes for specifying the maximum sizes
of partitions (the "partition limit"). This was inconsistent with the
rest of the FVM API which uses slices.

This converts the partition limit to slice units and updates the users.
The fshost configuration is left in bytes since the slice size isn't
necessarily available at that layer. fshost converts to slices when it
sets the limit.

Change-Id: I28e413dd5e99ad63b1c34992d5efe0cc4a3e8f8d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/606092
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
API-Review: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>