[fdf] Stop the driver component on Stop or Kill
When either ComponentController.Stop() or ComponentController.Kill() are
called, make sure we stop the driver component.

Previously, there was an ordering guarantee that Kill() would be called
after Stop(), and that at least Stop() would be issued. That is no
longer true, and though we set a stop-timeout for the environment that
the driver components are run in, this provides some additional caution.

This change also ensures that we correctly send an epitaph to the
component manager, so knows the component runner has stopped the
component.

Bug: 33183
Change-Id: Idaa660d00f7a0e3ff60778662413ce9be05682ed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/618005
Fuchsia-Auto-Submit: Abdulla Kamar <abdulla@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>