[driver manager] Clean up Coordinator code
Fixed some readability and style guide issues
in Coordinator:
- Removed extra include statements
- Rearranged header file so that the accessors and
  mutators are in one place, and the variables are
  defined after the functions
- Removed dead functions: InitInspect(),
  BuildSuspendList(), ValidateDrivers()

Moved the DFv1 driver initializing logic from
main.cc to a LoadDriversV1() function in Coordinator.
This lets us delete ScheduleBaseDriverLoading(),
BindDrivers(), and require_system(), reducing the
number of public functions that we need to keep
track of. It also reduces the risks of these
functions being used incorrectly, and puts
initialization into one place.

Modify InitCoreDevices() so it doesn't return a
zx_status. The function was not returning an error,
only ZX_OK, which was a misleading.

Bug: 89910
Change-Id: I7a61213accb432a69fdf4c9ce02ee9c0e1ddfa72
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/624181
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Sarah Chan <spqchan@google.com>