[x86][acpi][ec] Restore locking for non-global lock case.
While we don't need to acquire the global lock if _GL is not set,
synchronisation is still necessary when interacting with the EC IO
ports. This CL adds locking in cases where we don't use the global lock
to avoid races in access to the IO ports.

Bug: 90837
Test: confirm sensors work on Eve/Atlas.
Change-Id: Ib542ce952e7832a6a351013c573687cde6fe279d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/625802
Reviewed-by: Alex Legg <alexlegg@google.com>
Commit-Queue: Simon Shields <simonshields@google.com>