[fdf][compat] Rename parent_ to linked_device
The 'parent_' field actually refers to a linked device.
Two devices that are linked represent the same device that
spans different drivers.

A parent Driver will create a Device, then the DriverFramework
will create a new Driver to bind to the device. If the new
Driver is colocated with the old Driver, it will see the Device
in its Symbols. The new Driver will create its own Device object,
but link it with the other Device object that was created by
the parent Driver.

This naming should remove confusion around the field. This is
*not* a parent device.

All devices should go through `linked_device_` to access their
own fields (see how the Metadata functions work).

Bug: 90629

Change-Id: I5643bde4e94f1e34225ca6c9850e7e833d5f528a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/628461
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>