[build][fdf] Add build flag to enable dfv2
An build flag was added which allows us to add a GN argument which
enables DFv2. Eventually this will be turned on on a per board basis,
but the GN arg allows us to enable it for any board while we are
developing support for it.

It's worth noting that no board will successful boot at present if DFv2
is enabled.

Bug: 90539
Test: Add `enable_dfv2 = true` to //products/bringup.gni && fx qemu
Change-Id: I4f41a5c6436cadf1919590fb8102286a40cc40f7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/631265
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>