[driver_runtime] Convert dispatcher to use RefPtr.
When fdf_dispatcher_destroy() is called, the dispatcher
DispatchCallbacks callback may be currently running,
or can be the context from which fdf_dispatcher_destroy()
is called from. This change ensures that the fdf_dispatcher
object remains alive until the last callback returns.

This is still an issue for callbacks which happen without
the async loop.

---------------
Implementation
---------------
The initial RefPtr is created in fdf_dispatcher_create(),
and that reference is recovered in fdf_dispatcher_destroy().

Each time we call EventWaiter::BeginWaitWithRef(), a new
reference to the dispatcher is stored, which is then
passed to the Dispatcher::DispatchCallbacks function.

Multiply: fuchsia-pkg://fuchsia.com/driver_runtime_tests#meta/driver_runtime_tests.cm : 100

Bug: 87840
Change-Id: I32b6ccbb228568661f7ee54e78cd8a60367d4af4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635538
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>