[fdf][compat] Free devices before removing
When we are removing a Device object, call Release before
removing it.

This fixes LSAN leaks in integration tests with the compat shim,
as the DriverHost no longer leaks the DFv1 Device and Driver objects.

Also, in DFv2 all Devices are released, so release the test-parent
devices correctly.

Change-Id: I7dcb8ca9283d2e6e2367b0dc292be23470b048a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635583
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>