[fdf][compat] Prevent UAF in callback
We cannot pass raw pointers to FIDL callbacks, as we don't
have a guarantee that the pointer will be alive when the callback
is called.

Instead we should print the error in a future, as the future will
be cancelled if the class goes out of scope.
We have this guarantee because we wrap our futures in a 'scope' object.

Bug: 91222
Change-Id: Iaf8bb20a4e59c3f5575f1d888796db427e277f6d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640119
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>