[fdf] Set fallback in driver component manifest
Create an optional boolean field in driver component manifests called
"fallback". This will denote whether or not the driver is a fallback driver.
The default value is false. Also create an optional boolean parameter called
"fallback" for the GN template "fuchsia_driver_component". The template will
set the "fallback" field within the driver component manifest that it creates
to the parameter's value. The default value is false.

Bug: 92329
Change-Id: I50fad4cc2cb7afc7a56abfb6389bc4b03e8beaec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639086
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>