[ftl] Add a retry and some reporting
There is evidence that suggests we have prematurely stopped processing
the last map block due to an intermittent failure. To hopefully catch
this and the reason for failing, we will retry reading the last page and
see if we succeed on the second attempt. If we do, then we will report
that we succeeded (via Cobalt).

There is no easy way of testing this automatically, so I made a simple
tweak to simulate a failure and then verified that the metric turned up
via iquery. This won't appear in Cobalt until we update Lapis
configuration. This change is likely only temporary.

Bug: 87653
Change-Id: Ib6e796f2ab0a9b5fe28361938e88d3c04de9905b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639924
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Chris Suter <csuter@google.com>