[fdf][compat] Check if Node is valid after Bind
A node can be freed while we are waiting to see if it can be bound.
This can happen if the node's parent has decided to exit.

In fact, this has a chance at happening in any DFv1 compat test,
because the last driver of the stack will produce a node that
checks if it can be bound. It's possible for the test to finish
and start exiting before that last node is checked to be bound.

Multiply: compat-driver-test
Change-Id: I128868455b3a96ea6a40f2bc1ec6a04d39059e9c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640202
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>